notices-server/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
python-decouple==3.5
requests==2.26.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.7
uWSGI==2.0.20
Werkzeug==2.0.2
